feat: 체크 핸들러 meeting page에서 userItem으로 이동
- userItem check 상태 추가
- attendance check fetch 이동
- check 핸들러 이동
