+++ frontend/src/components/UserItem/index.tsx
@@ -1,5 +1,5 @@
 import { css } from '@emotion/react';
-import React from 'react';
+import React, { useState } from 'react';
 import * as S from './UserItem.styled';
 import CoffeeIconSVG from '../../assets/coffee.svg';
 import Checkbox from '../../components/@shared/Checkbox';
@@ -11,20 +11,54 @@ type User = {
   nickname: string;
   accessToken: null | string;
   tardyCount: number;
-  attendanceStatus: string;
+  attendanceStatus: 'present' | 'tardy';
 };
 
 type UserItemProps = {
   user: Omit<User, 'accessToken'>;
-  onAttendanceCheck: (
-    user: Omit<User, 'accessToken'>,
-    checked: boolean
-  ) => void;
+  meetingId: string;
 };
 
-const UserItem: React.FC<UserItemProps> = ({ user, onAttendanceCheck }) => {
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    onAttendanceCheck(user, e.target.checked);
+const ATTENDANCE_STATUS = {
+  tardy: false,
+  present: true,
+} as const;
+
+const userAttendanceFetch = async (url: string, payload: any) => {
+  return fetch(url, {
+    method: 'PUT',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify(payload),
+  });
+};
+
+const UserItem: React.FC<UserItemProps> = ({ user, meetingId }) => {
+  const [checked, setChecked] = useState<boolean>(
+    ATTENDANCE_STATUS[user.attendanceStatus]
+  );
+
+  const handleChange = async ({
+    target: { checked },
+  }: React.ChangeEvent<HTMLInputElement>) => {
+    setChecked(checked);
+
+    try {
+      const response = await userAttendanceFetch(
+        `/meetings/${meetingId}/users/${user.id}`,
+        {
+          attendanceStatus: checked ? 'present' : 'tardy',
+        }
+      );
+
+      if (!response.ok) {
+        setChecked(!checked);
+        alert('출석체크에 실패했습니다.');
+      }
+    } catch (error) {
+      alert('출석체크중 오류가 발생했습니다.');
+    }
   };
 
   return (
@@ -37,10 +71,7 @@ const UserItem: React.FC<UserItemProps> = ({ user, onAttendanceCheck }) => {
           ))}
         </S.CoffeeIconImageBox>
       </S.Box>
-      <Checkbox
-        onChange={handleChange}
-        checked={user.attendanceStatus === 'tardy' ? false : true}
-      />
+      <Checkbox onChange={handleChange} checked={checked} />
     </S.Layout>
   );
 };
